requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
pandas>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
albumentations>=1.4.0
geffnet>=1.0.2
timm>=0.9.0
scikit-learn>=1.3.0
pyyaml>=6.0
optuna>=3.5.0
mlflow>=2.10.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
kaleido>=0.2.1
python-dotenv>=1.0.0
